Task2/requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.2
opencv-python==4.4.0.42
Pillow==7.2.0
pytz==2020.1
six==1.15.0
torch==1.6.0
torchvision==0.7.0
Werkzeug==1.0.1
